forallx-yyc-{fol,tfl}.tex: Use 'name' consistently for constants in FOL
There were a few places where the text still referred to
'constants' (including the official definition of the FOL syntax in
Ch. 26) but without explanation of this terminology.  Since it appears
the intent was to use 'name' throughout (indicated especially by the
entry for 'Names' in appendix A), this commit changes the remaining
references to constants to use 'name' instead.