solutions/forallx-sol-prooffol.tex: fix solution to problem 35.A
Provide a justification for the fourth form of the rule.  The existing
proof actually just duplicated the justification for the first form of
the rule from the main text.  The fourth form of the rule is not
intuitionistically valid, and requires a more complicated proof.
I've added a proof that uses only basic rules.